+++ conlimiter/doc.go
@@ -1,9 +1,0 @@
-/*
-Package conlimiter provides a simple connection limiter for incoming TCP
-connections.
-
-Deprecated: conlimiter is no longer planned to be used by future WebPA/XMiDT services.
-
-This package is frozen and no new functionality will be added.
-*/
-package conlimiter