+++ store/doc.go
@@ -1,9 +1,0 @@
-/*
-Package store implements some additional atomic value storage on top of sync/atomic.
-In particular, this means transparent caching of arbitrary values.
-
-Deprecated: store is no longer planned to be used by future WebPA/XMiDT services.
-
-This package is frozen and no new functionality will be added.
-*/
-package store