+++ xwebhook/doc.go
@@ -1,8 +1,0 @@
-/*
-Package xwebhook provides an event webhook registry component for XMiDT services.
-
-Deprecated: xwebhook has moved to https://github.com/xmidt-org/ancla.
-
-This package is frozen and no new functionality will be added.
-*/
-package xwebhook